+++ test/fibers.jl
@@ -1,20 +1,10 @@
 using RecursiveArrayTools, ManifoldsBase, Test
 using Random
 using ManifoldsBase: DefaultManifold, VectorSpaceType, ℝ, Fiber
-struct TestVectorSpaceType <: VectorSpaceType end
 
-struct TestFiberType <: ManifoldsBase.FiberType end
-
-function ManifoldsBase.fiber_dimension(M::AbstractManifold, ::TestFiberType)
-    return 2 * manifold_dimension(M)
-end
-
-function ManifoldsBase.vector_space_dimension(M::AbstractManifold, ::TestVectorSpaceType)
-    return 2 * manifold_dimension(M)
-end
-
-include("test_manifolds.jl")
-
+s = @__DIR__
+!(s in LOAD_PATH) && (push!(LOAD_PATH, s))
+using ManifoldsBaseTestUtils
 
 @testset "vector space fibers" begin
     M = DefaultManifold(3)